refactor : LottoController 클래스 분리  - 자동 로또 구입처리를 담당하는 LottoAutoPurchaseController 클래스  - 로또 당첨결과 조회를 담당하는 LottoWinInquiryController 클래스
