+++ src/main/java/lotto/view/LottoAutoApplication.java
@@ -1,17 +1,19 @@
 package lotto.view;
 
-import lotto.controller.LottoController;
+import lotto.controller.LottoAutoPurchaseController;
+import lotto.controller.LottoWinInquiryController;
 import lotto.controller.dto.LottoAutoPurchaseResponse;
 
 public class LottoAutoApplication {
     public static void main(String[] args) {
         InputView inputView = new InputView();
         ResultView resultView = new ResultView();
-        LottoController lottoController = new LottoController();
+        LottoAutoPurchaseController lottoAutoPurchaseController = new LottoAutoPurchaseController();
+        LottoWinInquiryController lottoWinInquiryController = new LottoWinInquiryController();
 
-        LottoAutoPurchaseResponse lottoAutoPurchaseResponse = lottoController.purchaseLottoAutoTicket(inputView.inputPurchaseAmount());
+        LottoAutoPurchaseResponse lottoAutoPurchaseResponse = lottoAutoPurchaseController.purchaseLottoAutoTicket(inputView.inputPurchaseAmount());
         resultView.printPurchaseList(lottoAutoPurchaseResponse);
         System.out.println();
-        resultView.printLottoWinStatistic(lottoController.inquiryWin(inputView.inputWinNumber(lottoAutoPurchaseResponse.getLottoNumberList())));
+        resultView.printLottoWinStatistic(lottoWinInquiryController.inquiryWin(inputView.inputWinNumber(lottoAutoPurchaseResponse.getLottoNumberList())));
     }
 }
